pyproject.toml
[project]
name = "travel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=0.1.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "google-search-results>=2.4.2",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "serpapi>=0.1.5",
    "streamlit>=1.32.0",
]
